+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - NumberDocument.NumberRangeValidator bug fixed, values now included correctly in error message.
 - NumberField.valueRange(), minimumValue() and maximumValue() removed, accessible via builder, builder validation improved, related refactoring.
+- NumberField, setters removed, only used in tests, accessible via builder.
 ### is.codion.framework.domain
 - KeyGenerator.Identity removed.